+++ doc/api/es_bash_completion.sh
@@ -569,7 +569,7 @@ _es() {
 				COMPREPLY=( $(compgen -P "${cur%/*}" -W "/ /backup" -- "/${cur##*/}" ) )
 			fi
 			[[ "${action}" == "get" ]] && params="-full"
-			[[ "${action}" == "create" ]] || [[ "${action}" == "set" ]] && params="-status -is_compute -is_backup -owner -cpu_coef -ram_coef -note"
+			[[ "${action}" == "create" ]] || [[ "${action}" == "set" ]] && params="-status -is_compute -is_backup -owner -address -cpu_coef -ram_coef -note"
 		;;
 
 		/node/*/license|/node/*/license/)